Fix `Style/OneClassPerFile` cop - simple form components
